feat: added NavigationView and NavigationLink
In our FrontView.swift, I added NavigationView and changed
EntranceButton, which originally contained Button but now Navigation
Link. This link do one more job than just navigationing to next view,
which is to set isFirst true to false.
Have to add popup that can request user input, which would be the
intraID, and that should be registered to DB, in a new document named
user's UUID.